Museo/model/requirements.txt
Flask
pymongo
google-generativeai
folium
geopy
flask-cors
difflib
